Mermaid diagram support
